This project is for Brewed Awakenings
Brewed Awakenings is a local coffee shop that wants a web site that shows which products they are selling, and which employee is selling them. Here's all of the information they want on their web site.

A list of employees.
A list of products.
A list of orders that shows which product was sold, and which employee sold it.